tweak self-kill for "run locally"

+++ src/aws/bash/self-terminate.sh
@@ -7,14 +7,21 @@
 
 #
 # IMPORTANT - this script requires permissions
-#           - you can test your permissions with this command
+#           - you can (partially) test your permissions with this command
 #                 $ aws ec2 describe-instances
 #
 
 # NOTE: IMDSv2 requires a TOKEN to talk to metadata endpoint
 
-TOKEN=`curl -X PUT "http://169.254.169.254/latest/api/token" -H "X-aws-ec2-metadata-token-ttl-seconds: 21600"`
-INSTANCE_ID=`curl -H "X-aws-ec2-metadata-token: $TOKEN" http://169.254.169.254/latest/meta-data/instance-id` 
+echo Attempting to get TOKEN from metadata
 
-aws ec2 terminate-instances --instance-ids $INSTANCE_ID
+TOKEN=`curl -X PUT "http://169.254.169.254/latest/api/token" -H "X-aws-ec2-metadata-token-ttl-seconds: 21600"`
+if [  "x$TOKEN" != "x" ]
+then
+    INSTANCE_ID=`curl -H "X-aws-ec2-metadata-token: $TOKEN" http://169.254.169.254/latest/meta-data/instance-id` 
+
+    echo Attempting terminate-instances on $INSTANCE_ID
+    aws ec2 terminate-instances --instance-ids $INSTANCE_ID
+fi
+
